requirements.txt
Django==2.2.28
django-debug-toolbar==3.2.4
django-filter==2.4
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
pytils==0.3
python-dotenv==0.21.0
requests==2.26.0
sorl-thumbnail==12.7.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
Faker==12.0.1